"Add spawn locations to map structure and initialization"
This commit adds a new array called `spawn_locations` to the map_data structure in `map.c`. This array is populated during the initialization of a map. The `map_init` function now takes two parameters, a file handle to the map file and a file handle to the spawn locations file. It will read and assign the spawn locations to the new array in the map_data structure. This update allows the game to track the locations where entities can spawn on the map.